% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_raw.r
\docType{data}
\name{data_raw}
\alias{data_raw}
\title{Toy raw data that makes no sense}
\format{An object of class \code{data.frame} with 3406 rows and 54 columns.}
\usage{
data(data_raw)
}
\description{
Data arbitrarily created from existing data (contact Dan for additional details)
}
\examples{
library(predict2)
data(data_raw)
summary_check(data_raw)
}
\keyword{datasets}
