% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_predict2.r
\docType{data}
\name{lm_data}
\alias{lm_data}
\title{Toy data that makes no sense}
\format{An object of class \code{data.frame} with 50 rows and 2 columns.}
\usage{
data(lm_data)
}
\description{
Data arbitrarily created using:
\preformatted{
 set.seed(333)
 lm_data = setNames(as.data.frame(runif(50, min = 0, max = 50)), 'y')
 lm_data$x = lm_data$y + 1
}
}
\examples{
data(lm_data)
df_train = lm_data[1:25, ]
df_test = lm_data[26:50, ]

lm_int = lm(y ~ x, data = df_train)
lm_noint = lm(y ~ 0 + x, data = df_train)
list_lm = c('lm_int', 'lm_noint')

predicted = predict2(
  model = list_lm,
  newdata = df_test,
  actual = df_test$y,
  pred_type = 'response',
  write_model = FALSE,
  write_pred = FALSE,
  csv_name = 'none'
)
head(predicted)
}
\keyword{datasets}
