% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sales.r
\docType{data}
\name{sales}
\alias{sales}
\title{Toy data that makes no sense}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5181 rows and 54 columns.}
\usage{
data(sales)
}
\description{
Data arbitrarily created from existing data (contact Dan for additional details)
}
\examples{
library(predict2)
data(sales)
subset(trfm_data, keep == 'y')
}
\keyword{datasets}
