% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_summary_check.R
\name{summary_check}
\alias{summary_check}
\title{Takes predictions from model object(s) and and combines it with actual values
This function takes a vector of strings that represent the model object names}
\usage{
summary_check(generic_df)
}
\arguments{
\item{generic_df}{The dataframe or datatable from which to get summarized information}
}
\value{
summary_check() This function will return the following moments for numeric columns of the data specified:
  mean, media, NA count, total rows, ratio of NA to total rows
}
\description{
Takes predictions from model object(s) and and combines it with actual values
This function takes a vector of strings that represent the model object names
}
