% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_trfm.r
\name{trfm}
\alias{trfm}
\title{Takes predictions from model object(s) and and combines it with actual values
This function takes a vector of strings that represent the model object names
trfm <- function(df_variable, missing, min, max, zero_value, transformation)}
\usage{
trfm(df_variable, missing, min, max, zero_value, transformation)
}
\arguments{
\item{df_variable}{A vector that contains the original values that need to be transformed}

\item{missing}{The value that should replace any  missing values in the dataset}

\item{min}{The minimum value that a given variable should take; any value below the minimum will be set equal to the min value}

\item{max}{The maximum value that a given variable should take; any value above the maximum will be set equal to the max value}

\item{zero_value}{This is the value that any value of 0 should be.}

\item{transformation}{This is an integer value for corresponding to one of three transformation: 0 == no transformation, 1 == log(), 2 == sqrt()}
}
\value{
trfm() will a vector of transformed values
}
\description{
Takes predictions from model object(s) and and combines it with actual values
This function takes a vector of strings that represent the model object names
trfm <- function(df_variable, missing, min, max, zero_value, transformation)
}
