% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trfm.r
\docType{data}
\name{trfm_data}
\alias{trfm_data}
\title{Toy data that makes no sense}
\format{An object of class \code{data.frame} with 49 rows and 7 columns.}
\usage{
data(trfm_data)
}
\description{
Data arbitrarily created from existing data (contact Dan for additional details)
}
\examples{
library(predict2)
data(trfm_data)
subset(trfm_data, keep == 'y')
}
\keyword{datasets}
