% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_twec_fiscal_month.R
\name{twec_fiscal_month}
\alias{twec_fiscal_month}
\title{Takes predictions from model object(s) and and combines it with actual values
This function takes a vector of strings that represent the model object names}
\usage{
twec_fiscal_month(month)
}
\arguments{
\item{month}{A vector or data column containig date (either as a Date class or character). Must be in the format of YYYY-MM-DD (either as.Date format or character)}
}
\value{
twec_fiscal_month() will return a vector of integers representing the fiscal year month designations for TWEC
}
\description{
Takes predictions from model object(s) and and combines it with actual values
This function takes a vector of strings that represent the model object names
}
\examples{
library(predict2)

# example with vector containing character
date_vec = c('2017-01-30', '2017-12-21')
twec_fiscal_month(date_vec)

# example with a data frame already as.Date format
date_df = setNames(data.frame(c('2017-01-30', '2017-04-05', '2017-12-26')), c('date'))
date_df$date = as.Date(date_df$date, format = '\%Y-\%m-\%d')
str(date_df)
twec_fiscal_month(date_df$date)
}
